Refactor: create package "data"
move packages: "content_provider", "database", and "network" into the data package
