+++ app/src/main/java/org/kiwix/kiwixmobile/data/database/entity/RecentSearchSpec.java
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
-package org.kiwix.kiwixmobile.database.entity;
+package org.kiwix.kiwixmobile.data.database.entity;
 
 import com.yahoo.squidb.annotations.ColumnSpec;
 import com.yahoo.squidb.annotations.TableModelSpec;
