+++ app/src/main/java/org/kiwix/kiwixmobile/data/network/KiwixService.java
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
-package org.kiwix.kiwixmobile.network;
+package org.kiwix.kiwixmobile.data.network;
 
 import org.kiwix.kiwixmobile.library.entity.LibraryNetworkEntity;
 import org.kiwix.kiwixmobile.library.entity.MetaLinkNetworkEntity;
