+++ app/src/main/java/org/kiwix/kiwixmobile/data/network/UserAgentInterceptor.java
@@ -15,21 +15,29 @@
  * You should have received a copy of the GNU General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
-package org.kiwix.kiwixmobile.database.entity;
+package org.kiwix.kiwixmobile.data.network;
 
-import com.yahoo.squidb.annotations.ColumnSpec;
-import com.yahoo.squidb.annotations.TableModelSpec;
+import java.io.IOException;
+
+import okhttp3.Interceptor;
+import okhttp3.Request;
+import okhttp3.Response;
 
 /**
- * Squidb spec for saved bookmarks.
+ * Created by mhutti1 on 20/04/17.
  */
-@TableModelSpec(className = "Bookmarks", tableName = "Bookmarks")
-public class BookmarksSpec {
 
-  @ColumnSpec(constraints = "NOT NULL")
-  public String ZimId;
-  public String ZimName;
-  public String bookmarkUrl;
-  public String bookmarkTitle;
+public class UserAgentInterceptor implements Interceptor{
+  public final String useragent;
 
+  public UserAgentInterceptor(String useragent) {
+    this.useragent = useragent;
+  }
+
+  @Override
+  public Response intercept(Chain chain) throws IOException {
+    Request originalRequest = chain.request();
+    Request newUserAgent = originalRequest.newBuilder().header("User-Agent", useragent).build();
+    return chain.proceed(newUserAgent);
+  }
 }
