+++ app/src/main/java/org/kiwix/kiwixmobile/di/modules/NetworkModule.java
@@ -21,8 +21,8 @@ import android.content.Context;
 import android.net.ConnectivityManager;
 
 import org.kiwix.kiwixmobile.BuildConfig;
-import org.kiwix.kiwixmobile.network.KiwixService;
-import org.kiwix.kiwixmobile.network.UserAgentInterceptor;
+import org.kiwix.kiwixmobile.data.network.KiwixService;
+import org.kiwix.kiwixmobile.data.network.UserAgentInterceptor;
 
 import javax.inject.Singleton;
