+++ app/src/main/java/org/kiwix/kiwixmobile/zim_manager/library_view/LibraryPresenter.java
@@ -20,10 +20,10 @@ package org.kiwix.kiwixmobile.zim_manager.library_view;
 import android.util.Log;
 
 import org.kiwix.kiwixmobile.base.BasePresenter;
-import org.kiwix.kiwixmobile.database.BookDao;
+import org.kiwix.kiwixmobile.data.database.BookDao;
 import org.kiwix.kiwixmobile.downloader.DownloadFragment;
 import org.kiwix.kiwixmobile.library.entity.LibraryNetworkEntity;
-import org.kiwix.kiwixmobile.network.KiwixService;
+import org.kiwix.kiwixmobile.data.network.KiwixService;
 import org.kiwix.kiwixmobile.zim_manager.library_view.contract.LibraryViewCallback;
 
 import javax.inject.Inject;
